add Wall files and rename Textures
