+++ C/raycast/raycast/Wall.h
@@ -1,0 +1,14 @@
+#ifndef WALL_H
+#define WALL_H
+
+#include <math.h>
+
+#include "Definitions.h"
+#include "Ray.h"
+#include "Player.h"
+#include "Graphics.h"
+#include "Texture.h"
+
+void RenderWallProjection(void);
+
+#endif // !WALL_H
